Phase by eaactid and worktype - ENDPOINT ADDED
