+++ src/reports_api/resources/__init__.py
@@ -20,6 +20,7 @@ All services have 2 defaults sets of endpoints:
  - meta
 That are used to expose operational health information about the service, and meta information.
 """
+from importlib.resources import path
 from flask import Blueprint
 
 from .code import API as CODES_API
@@ -27,6 +28,7 @@ from .ops import API as OPS_API
 from .apihelper import Api
 from .meta import API as META_API
 from .project import API as PROJECTS_API
+from .phase import API as PHASE_API
 
 __all__ = ('API_BLUEPRINT', 'OPS_BLUEPRINT')
 
@@ -60,3 +62,4 @@ API = Api(
 API.add_namespace(META_API, path='/meta')
 API.add_namespace(CODES_API, path='/codes')
 API.add_namespace(PROJECTS_API, path='/projects')
+API.add_namespace(PHASE_API, path='/phases')
